Extract actions grid component
