+++ packages/app/src/features/actions/ActionsContainer.tsx
@@ -6,13 +6,13 @@ import { stringifyObjectivesToStableActions } from './logic/stringifyObjectives'
 import { InjectedActionsContext, Objective } from './logic/types'
 import { useActionHandlers } from './logic/useActionHandlers'
 import { useSettingsDialog } from './settings-dialog/logic/useSettingsDialog'
-import { ActionsView } from './views/ActionsView'
+import { ActionsView, ActionsViewVariant } from './views/ActionsView'
 
 export interface ActionsContainerProps {
   objectives: Objective[]
   context?: InjectedActionsContext
   onFinish?: () => void // called only once, after render when all actions are marked successful
-  variant?: 'default' | 'dialog'
+  variant?: ActionsViewVariant
   enabled?: boolean
 }
 
@@ -20,7 +20,7 @@ function ActionsContainer({
   objectives,
   context,
   onFinish,
-  variant = 'default',
+  variant = 'extended',
   enabled,
 }: RequireKeys<ActionsContainerProps, 'enabled'>) {
   const { handlers, settingsDisabled } = useActionHandlers(objectives, {
